Keep better track of rejected rules. (#34)
This PR writes rejected rules into a file and checks against this file
each compile. This allows us to review for new rules that fail using git
diff and supporess error messages for rules which are already known to
be broken.

This PR also returns a failed error code when an unhandled error occurs
which allows it to be effective in CI now.

Additionally this PR also started adding some VQL based sigma rules for
example checking for presence of files or registry keys.