Updated Linux EBPF model to fall back to event process info
if the process tracker is not initialized then it may not be able to
find the relevant process (if it is very short lived) but in this case
some of the information is already delivered in the event
itself. Copying it into the ProcInfo field will allow rules to match
against these.